connect: carve push.rs (pure move)
Web Push leaves main.rs: the RFC 8291 payload encryption and RFC 8292
VAPID signing (pure ring), the subscribe/unsubscribe/test API, and the
presence-alert and handle-reclaim monitors that author the alerts,
plus the webpush/VAPID tests.
